Add files via upload
Added a link to a folder on Google Drive with code on Visual Studio Community that are answers to quizzes on learncpp.com. It is tedious to create the same file structure on GitHub, and the code is just to demonstrate my experience with learning C++, rather than any attempt at a collaborative effort.